v1.0.6
🚀 Refactor: Refactor `NetworkPrinter` into `GenericPrinter` class.
🚀 New `BytesPrinter`.

- sdk: ^3.6.0

- resource_portable: ^3.1.2

- test: ^1.25.15
- coverage: ^1.11.1
- intl: ^0.20.2
