+++ lib/src/printer/network_printer.dart
@@ -11,31 +11,12 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data' show Uint8List;
 
-import 'package:image/image.dart';
-
-import '../utils/barcode.dart';
-import '../utils/capability_profile.dart';
-import '../utils/enums.dart';
-import '../utils/generator.dart';
-import '../utils/pos_column.dart';
-import '../utils/pos_styles.dart';
-import '../utils/qrcode.dart';
 import 'enums.dart';
+import 'generic_printer.dart';
 
-/// Network Printer
-class NetworkPrinter {
-  final PaperSize _paperSize;
-  final CapabilityProfile _profile;
-
-  late final Generator _generator;
-
-  NetworkPrinter(this._paperSize, this._profile, {int spaceBetweenRows = 5})
-      : _generator =
-            Generator(_paperSize, _profile, spaceBetweenRows: spaceBetweenRows);
-
-  PaperSize get paperSize => _paperSize;
-
-  CapabilityProfile get profile => _profile;
+/// Network ESC/POS Printer.
+class NetworkPrinter extends GenericPrinter {
+  NetworkPrinter(super._paperSize, super._profile, {super.spaceBetweenRows});
 
   String? _host;
 
@@ -56,7 +37,6 @@ class NetworkPrinter {
       {int port = 91000, Duration timeout = const Duration(seconds: 5)}) async {
     _host = host;
     _port = port;
-
     return await ensureConnected(timeout: timeout);
   }
 
@@ -78,7 +58,7 @@ class NetworkPrinter {
       _connected = true;
 
       _socket.listen(_addInputBytes);
-      _socket.add(_generator.reset());
+      _socket.add(generator.reset());
 
       return Future<PosPrintResult>.value(PosPrintResult.success);
     } catch (e) {
@@ -97,6 +77,9 @@ class NetworkPrinter {
     _socket.destroy();
     _disposeInputBytes();
   }
+
+  @override
+  void writeBytes(List<int> bytes) => _socket.add(bytes);
 
   void _disposeInputBytes() {
     _inputBytes.clear();
@@ -135,146 +118,10 @@ class NetworkPrinter {
     return future;
   }
 
-  // ************************ Printer Commands ************************
-  void reset() {
-    _socket.add(_generator.reset());
-  }
-
-  void endJob() {
-    _socket.add(_generator.endJob());
-  }
-
-  void text(
-    String text, {
-    PosStyles styles = const PosStyles(),
-    int linesAfter = 0,
-    bool containsChinese = false,
-    int? maxCharsPerLine,
-  }) {
-    _socket.add(_generator.text(text,
-        styles: styles,
-        linesAfter: linesAfter,
-        containsChinese: containsChinese,
-        maxCharsPerLine: maxCharsPerLine));
-  }
-
-  void setGlobalCodeTable(String codeTable) {
-    _socket.add(_generator.setGlobalCodeTable(codeTable));
-  }
-
-  void setGlobalFont(PosFontType font, {int? maxCharsPerLine}) {
-    _socket
-        .add(_generator.setGlobalFont(font, maxCharsPerLine: maxCharsPerLine));
-  }
-
-  void setStyles(PosStyles styles, {bool isKanji = false}) {
-    _socket.add(_generator.setStyles(styles, isKanji: isKanji));
-  }
-
-  void rawBytes(List<int> cmd, {bool isKanji = false}) {
-    _socket.add(_generator.rawBytes(cmd, isKanji: isKanji));
-  }
-
-  void emptyLines(int n) {
-    _socket.add(_generator.emptyLines(n));
-  }
-
-  void feed(int n) {
-    _socket.add(_generator.feed(n));
-  }
-
-  void cut({PosCutMode mode = PosCutMode.full}) {
-    _socket.add(_generator.cut(mode: mode));
-  }
-
-  void printCodeTable({String? codeTable}) {
-    _socket.add(_generator.printCodeTable(codeTable: codeTable));
-  }
-
-  void beep({int n = 3, PosBeepDuration duration = PosBeepDuration.beep450ms}) {
-    _socket.add(_generator.beep(n: n, duration: duration));
-  }
-
-  void reverseFeed(int n) {
-    _socket.add(_generator.reverseFeed(n));
-  }
-
-  void row(List<PosColumn> cols) {
-    _socket.add(_generator.row(cols));
-  }
-
-  void image(Image imgSrc, {PosAlign align = PosAlign.center}) {
-    _socket.add(_generator.image(imgSrc, align: align));
-  }
-
-  void imageRaster(
-    Image image, {
-    PosAlign align = PosAlign.center,
-    bool highDensityHorizontal = true,
-    bool highDensityVertical = true,
-    PosImageFn imageFn = PosImageFn.bitImageRaster,
-  }) {
-    _socket.add(_generator.imageRaster(
-      image,
-      align: align,
-      highDensityHorizontal: highDensityHorizontal,
-      highDensityVertical: highDensityVertical,
-      imageFn: imageFn,
-    ));
-  }
-
-  void barcode(
-    Barcode barcode, {
-    int? width,
-    int? height,
-    BarcodeFont? font,
-    BarcodeText textPos = BarcodeText.below,
-    PosAlign align = PosAlign.center,
-  }) {
-    _socket.add(_generator.barcode(
-      barcode,
-      width: width,
-      height: height,
-      font: font,
-      textPos: textPos,
-      align: align,
-    ));
-  }
-
-  void qrcode(
-    String text, {
-    PosAlign align = PosAlign.center,
-    QRSize size = QRSize.size4,
-    QRCorrection cor = QRCorrection.L,
-  }) {
-    _socket.add(_generator.qrcode(text, align: align, size: size, cor: cor));
-  }
-
-  void drawer({PosDrawer pin = PosDrawer.pin2}) {
-    _socket.add(_generator.drawer(pin: pin));
-  }
-
-  void hr({String ch = '-', int? len, int linesAfter = 0}) {
-    _socket.add(_generator.hr(ch: ch, linesAfter: linesAfter));
-  }
-
-  void textEncoded(
-    Uint8List textBytes, {
-    PosStyles styles = const PosStyles(),
-    int linesAfter = 0,
-    int? maxCharsPerLine,
-  }) {
-    _socket.add(_generator.textEncoded(
-      textBytes,
-      styles: styles,
-      linesAfter: linesAfter,
-      maxCharsPerLine: maxCharsPerLine,
-    ));
-  }
-
   Future<int?> transmissionOfStatus({int n = 1}) async {
     var waitFuture = _waitInputByte();
-    _socket.add(_generator.transmissionOfStatus(n: n));
+    writeBytes(generator.transmissionOfStatus(n: n));
+
     await waitFuture;
 
     var status = _inputBytes.lastOrNull;
@@ -284,6 +131,4 @@ class NetworkPrinter {
     }
     return status;
   }
-
-// ************************ (end) Printer Commands ************************
 }
