+++ pubspec.yaml
@@ -1,21 +1,21 @@
 name: esc_pos_dart
 description: ESC/POS thermal WiFi printer support for Dart (non-Flutter dependent).
-version: 1.0.5
+version: 1.0.6
 repository: https://github.com/gmpassos/esc_pos_dart
 
 environment:
-  sdk: ^3.4.0
+  sdk: ^3.6.0
 
 dependencies:
   hex: ^0.2.0
   gbk_codec: ^0.4.0
   image: ^3.3.0
-  resource_portable: ^3.1.0
+  resource_portable: ^3.1.2
   collection: ^1.19.1
 
 dev_dependencies:
   lints: ^4.0.0
-  test: ^1.25.11
+  test: ^1.25.15
   dependency_validator: ^3.2.3
-  coverage: ^1.11.0
-  intl: ^0.20.1
+  coverage: ^1.11.1
+  intl: ^0.20.2
